+++ tests/benchmark.py
@@ -4,8 +4,9 @@ import os
 import tempfile
 import timeit
 
+from collections.abc import Callable
 from dataclasses import asdict, dataclass
-from typing import Callable, List, Optional, Tuple, Union
+from typing import List, Optional, Tuple
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -58,7 +59,7 @@ def read_files_serial(file_paths: List[str]) -> None:
 class BenchMark:
     """A class to benchmark the performance of the system."""
 
-    def __init__(self, name: str, show_graphs: bool = False):
+    def __init__(self, name: str, show_graphs: bool = False) -> None:
         """Create a benchmark object to measure the performance of the system."""
         self.results: Optional[Performance] = None
         self.graphs: bool = show_graphs
@@ -72,7 +73,7 @@ class BenchMark:
         file_counts: NDArray[np.float64],
         time_values: NDArray[np.float64],
         z_axis_label: str,
-        scatter_colors: Union[str, NDArray[np.int64]],
+        scatter_colors: str | NDArray[np.int64],
         view_tuple: Tuple[int, int],
     ) -> None:
         """Plot a mesh of values on a scatter plot graph.
@@ -251,7 +252,7 @@ class BenchMark:
             )
 
 
-def run_benchmark():
+def run_benchmark() -> None:
     """Run the benchmark tests for both parallel and serial file read/write methods."""
     benchmark_parallel = BenchMark("Parallel")
     benchmark_parallel.measure_times(